refactor: extract ensureMeshCoreManagerStarted helper; collapse double-filter
- meshcoreConfig.ts: add ensureMeshCoreManagerStarted(source, cfg) — the
  canonical create-or-connect recipe (create if missing, reconnect if
  disconnected, skip+log if already connected)
- sourceRoutes.ts: replace 4 inline create-or-connect blocks with the helper;
  the /connect endpoint with alreadyRunning logic is kept at its call site
- server.ts: startup auto-connect loop also replaced with the helper; removes
  the now-redundant MeshCoreManager import from server.ts
- messageRoutes.ts + v1/messages.ts: collapse .filter(isMeshCoreManager)
  .filter(m => m.isConnected()) into a single type-predicate filter
